# lib/api-zod/src/generated/types/bidderProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bidder Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BidderProfile {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  photoObjectPath?: string | null;
  /** @nullable */
  resumeObjectPath?: string | null;
  /** @nullable */
  resumeFileName?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  experience?: string | null;
  createdAt: string;
  updatedAt: string;
}