# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bidder Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  reportId: number;
  authorId: number;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorRole?: string | null;
  content: string;
  createdAt: string;
}