# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bidder Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  senderId: number;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderRole?: string | null;
  content: string;
  createdAt: string;
}